+++ models/User.js
@@ -1,5 +1,4 @@
 const mongoose = require('mongoose');
-
 const UserSchema = new mongoose.Schema({
 	firstName: {
 		type: String,
