+++ package.json
@@ -8,20 +8,21 @@
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "bl": ">=2.2.1",
     "body-parser": "^1.19.0",
     "connect-flash": "^0.1.1",
     "connect-mongo": "^3.2.0",
     "cookie-parser": "^1.4.5",
-    "ejs": "^3.1.3",
+    "ejs": "^3.1.5",
     "eslint-config-airbnb": "^18.2.0",
-    "eslint-plugin-import": "^2.21.2",
+    "eslint-plugin-import": "^2.22.0",
     "express": "^4.17.1",
     "express-ejs-layouts": "^2.5.0",
     "express-session": "^1.17.1",
     "husky": "^4.2.5",
     "jsonwebtoken": "^8.5.1",
     "latest": "^0.2.0",
-    "mongoose": "^5.9.27",
+    "mongoose": "^5.10.2",
     "n": "^6.7.0",
     "node-json-db": "^1.1.0",
     "nodemon": "^2.0.4",
@@ -29,9 +30,9 @@
     "passport-local": "^1.0.0"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.5",
-    "eslint": "^7.5.0",
-    "jest": "^26.1.0"
+    "@types/jest": "^26.0.13",
+    "eslint": "^7.8.1",
+    "jest": "^26.4.2"
   },
   "scripts": {
     "dev": "nodemon app.js -e js",
