--- public/.eslintrc.json
@@ -1,11 +1,0 @@
-{
-	"env": {
-		"commonjs": true,
-		"es2020": true,
-		"browser": true
-	},
-	"rules": {
-		"no-restricted-globals": "off",
-		"no-console": "off"
-	}
-}
