+++ routes/users.js
@@ -1,102 +1,23 @@
 const express = require('express');
 const router = express.Router();
-const bcrypt = require('bcryptjs');
-const passport = require('passport');
+const UsersController = require('../controllers/users');
 
-// Load User model
-const User = require('../models/User');
 
 // Login Page
-router.get('/login', (req, res) => res.render('login',{title:'Login'}));
+router.get('/login', UsersController.getLoginPage );
 
 
 // Register Page
-router.get('/register', (req, res) => res.render('register',{title:'Register'}));
+router.get('/register', UsersController.getRegisterPage );
 
 // Register
-router.post('/register', (req, res) => {
-	const { firstName, lastName, email, password, password2 } = req.body;
-	let errors = [];
-
-	if (!firstName || !lastName || !email || !password || !password2) {
-		errors.push({ msg: 'Please enter all fields' });
-	}
-
-	if (password != password2) {
-		errors.push({ msg: 'Passwords do not match' });
-	}
-
-	if (password.length < 6) {
-		errors.push({ msg: 'Password must be at least 6 characters' });
-	}
-
-	if (errors.length > 0) {
-		res.render('register', {
-			title:'Register',
-			errors,
-			firstName,
-			lastName,
-			email,
-			password,
-			password2
-		});
-	} else {
-		User.findOne({ email: email }).then(user => {
-			if (user) {
-				errors.push({ msg: 'Email already exists' });
-				res.render('register', {
-					title:'Register',
-					errors,
-					firstName,
-					lastName,
-					email,
-					password,
-					password2
-				});
-			} else {
-				const newUser = new User({
-					firstName,
-					lastName,
-					email,
-					password
-				});
-
-				bcrypt.genSalt(10, (err, salt) => {
-					bcrypt.hash(newUser.password, salt, (err, hash) => {
-						if (err) throw err;
-						newUser.password = hash;
-						newUser
-							.save()
-							.then(() => {
-								req.flash(
-									'success_msg',
-									'You are now registered and can log in'
-								);
-								res.redirect('/users/login');
-							})
-							.catch(err => console.log(err));
-					});
-				});
-			}
-		});
-	}
-});
+router.post('/register', UsersController.registerUser);
 
 //login handle
-router.post('/login', (req, res, next) => {
-	passport.authenticate('local', {
-		successRedirect: '/dashboard',
-		failureRedirect: '/users/login',
-		failureFlash: true
-	})(req, res, next);
-});
+router.post('/login', UsersController.handleLogin);
 
 //logout handle
-router.get('/logout', (req, res) => {
-	req.logOut();
-	req.flash('success_msg', 'You are logged out');
-	res.redirect('/users/login');
-});
+router.get('/logout', UsersController.handleLogout);
 
 
 module.exports = router;
